[feature] fusepy: Pass flags to create in non raw_fi mode.
fusepy/fusepy 127

Calls to open with O_CREAT on a file system will end up with a call to
create on FUSE. So it is possible to "create" with O_RDWR. Users need
to make the distinction between O_RDWR and O_WRONLY.

It is common to open with O_CREAT|O_RDWR before mmap'ing the new file.
If the filesystem does not allow reading of that file, it will cause a
bus error.