[performance] Mount TARs recursively after analyzing the outer TAR
This way frequent backward seeks are avoided and creating the index for
xz files was sped up from ~110s to 1-2s!

Note that the order of adding files to the SQLite database is not
important anymore. Duplicate results are sorted by offset anyways.

The only problem I see with this, is that in the case of a TAR
containing millions of TARs, the memory for the temporarily saved
FileInfo objects, might lead to unbounded memory usage.