[fix] Resolve symbolic links pointing outside the source folder to not break them
I see no reason to implement this logic for the other mount sources
because the links are not expected to work correctly anyway when they
point outside of the archive.

This logic will bring its own troubles in case this happens for a write
overlay! E.g., writing into a folder that actually is a resolved
symbolic link will probably fail because the folder cannot be created
ontop of that link. Or it might just work and write into the pointed to
folder.