[build] AppImage: Source the Python-AppImage Python wrapper script to avoid invoking bash after changing LD_LIBRARY_PATH
So many hours, which could have been spent on new features or
performance improvements, were lost in trying to fix these few lines of
incantation to avoid the bash warning about:

    libtinfo.so.6: no version information available

I do not even know where to file a bug. It worked with manylinux2014, so
maybe file a bug with manylinux? Or is AlmaLinux 8 at fault, or
libcurses for bundling some of its shared libraries without version
information?