[performance] Forward underlying archive block sizes to statfs and stat implementations
This is especially important for Lustre after disabling the buffering
for calls through FUSE in the earlier commit. Without this, we would now
have 8K reads to Lustre that has huge latencies and should function more
optimally with 4 MiB reads.

See also the proposal to increase the default buffer size of 8K:
https://github.com/python/cpython/issues/117151