[performance] Disable buffering when opening files via FUSE
Buffering should be done by the caller of the FUSE layer, which can be
anything from an Assember to Python program.
Furthermore, buffering may also be done on other layers, e.g., the file
system / page cache the archive resides on.

This partly fixes write amplification for archives residing on a Lustre
file system with 4 MiB advertized block size used for the default buffer
size. When accessing any granularity < 4 MiB, this leads to data being
read that was not necessary.