[doc] Remove redundant SQL benchmark results
Note that these data files contained timing for each insert when
inserting n rows, i.e., the benchmarks for up to 10M rows will include
the results for 100k rows and so on in its first 100k data points.

The results for 2M was kept because it was the largest one for which the
plots also tried to scale the SELECT statement, which is always the same
because the SELECT statements are repeated on the completed database!

16M was removed because it did not have a corresponding plot.