[style] Remove dead branch for parallel TAR entry processing
It was hidden because it slows things down in certain circumstances.
But, I did observe a nice speedup for TARs containing only empty files.
Probably unchanged for files <= 512B but for anything else it is doubtful
whether this parallelization is helpful because of the high amount of
random access necessary.

 - Parallelizing is only possible for actual files not objects because
   the file needs to be reopened from a different process
   (not just thread to circumvent the global interpreter lock).
 - For compressed files, the bottleneck is the decompression and the
   necessary seeks for this parallelization scheme would slow that down further.
 - Did only work on Linux because of multi-file access problems on macOS and Windows.