[fix] Make @overrides a test-time error not a runtime-error
The type-checking is generally too strict. Often it will also work with
non-matching types, e.g., if fusepy returns float times instead of int
or a List instead of a Tuple. It makes no sense to make this a runtime
error and risk making type-hint changes upstream a breaking chance!