[refactor] Remove unnecessary seek
This seeks back after recursive mounting, i.e., for recursion the
streaming mode with 'r|' was useless anyway!
It is not necessary because CPython tarfile stores the offset in
TarFile.offset and seeks to it before reading the next TarInfo.
Even if it was necessary, it should have been done more centrally
inside _TarFileMetadataReader.process.